cleaned up `./api` and made it consume `models`, `utils` & `tooling`
