--- api/.eslintignore
@@ -1,11 +1,0 @@
-node_modules/
-coverage
-dist
-src/_common
-oracle-cloud/build
-*.*
-!*.tsx
-!*.ts
-!*.js
-!*.json
-*.svg
