--- api/.eslintrc.json
@@ -1,29 +1,0 @@
-{
-  "root": true,
-  "parser": "@typescript-eslint/parser", // Specifies the ESLint parser
-  "extends": [
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended", // Uses the recommended rules from the @typescript-eslint/eslint-plugin
-    "prettier" // Turns off all rules that are unnecessary or might conflict with Prettier
-  ],
-  "parserOptions": {
-    "ecmaVersion": 2018, // Allows for the parsing of modern ECMAScript features
-    "sourceType": "module" // Allows for the use of imports
-  },
-  "plugins": ["@typescript-eslint", "unicorn"],
-  "rules": {
-    "sort-imports": [
-      "warn",
-      {
-        "ignoreCase": false,
-        "ignoreDeclarationSort": false,
-        "ignoreMemberSort": false,
-        "memberSyntaxSortOrder": ["none", "all", "multiple", "single"]
-      }
-    ],
-    "unicorn/filename-case": "error",
-    "@typescript-eslint/explicit-function-return-type": "off",
-    "@typescript-eslint/explicit-module-boundary-types": "off",
-    "camelcase": 1
-  }
-}
