--- api/.gitignore
@@ -1,11 +1,0 @@
-dist
-src/_common
-coverage
-oracle-cloud/build
-fetch_cache
-
-# logs
-**.tsbuildinfo
-
-# node
-**node_modules
