--- api/.prettierignore
@@ -1,13 +1,0 @@
-node_modules/
-coverage
-dist
-src/_common
-oracle-cloud/build
-
-.*
-*.tsbuildinfo
-*.log
-*.svg
-*.jpeg
-*.png
-*.ico
