--- api/.prettierrc
@@ -1,6 +1,0 @@
-{
-  "singleQuote": false,
-  "printWidth": 100,
-  "trailingComma": "all",
-  "endOfLine": "auto"
-}
