+++ api/src/contribution/repository.ts
@@ -1,22 +1,22 @@
-import { FilterDto, OptionDto } from "../_common/api/responses";
-import { ContributionEntity } from "../_common/entities/contribution";
-import { GetContributionsResponseDto } from "../_common/api/responses";
 import { GithubService } from "../github/service";
-import { Model } from "../_common/entities";
-import { ProjectReferenceEntity } from "../_common/entities/project-reference";
 import { Service } from "typedi";
-import { getDataCollection } from "../_common/utils/data";
-import { join } from "path";
+// import { getDataCollection } from "../_common/utils/data";
+// import { join } from "path";
+import { Model } from "@dzcode.io/models/dist/_base";
+import { ProjectReferenceEntity } from "@dzcode.io/models/dist/project-reference";
+import { ContributionEntity } from "@dzcode.io/models/dist/contribution";
+import { FilterDto, GetContributionsResponseDto, OptionDto } from "./types";
 
 @Service()
 export class ContributionRepository {
   constructor(private readonly githubService: GithubService) {
-    const projects = getDataCollection<Model<ProjectReferenceEntity, "repositories">>(
-      join(__dirname, "../../../../data"),
-      "projects-v2",
-      "list.json",
-    );
-    this.projects = projects !== 404 ? projects : [];
+    // const projects = getDataCollection<Model<ProjectReferenceEntity, "repositories">>(
+    //   join(__dirname, "../../../../data"),
+    //   "projects-v2",
+    //   "list.json",
+    // );
+    // this.projects = projects !== 404 ? projects : [];
+    this.projects = [];
   }
 
   private projects: Model<ProjectReferenceEntity, "repositories">[];
@@ -42,7 +42,7 @@ export class ContributionRepository {
                   repo,
                 });
                 return issuesIncludingPRs.map<Model<ContributionEntity, "project">>(
-                  /* eslint-disable camelcase */
+                   
                   ({
                     number,
                     labels: gLabels,
