+++ api/src/contributor/controller.ts
@@ -1,8 +1,9 @@
 import { Controller, Get, QueryParam } from "routing-controllers";
-import { GetContributorsResponseDto, GithubUserDto } from "../_common/api/responses";
 import { OpenAPI, ResponseSchema } from "routing-controllers-openapi";
 import { GithubService } from "../github/service";
 import { Service } from "typedi";
+import { GetContributorsResponseDto } from "./types";
+import { GithubUserDto } from "../github-user/types";
 
 @Service()
 @Controller("/Contributors")
