+++ api/src/github/service.spec.ts
@@ -2,8 +2,8 @@ import { GeneralGithubQuery, ListContributorsResponse } from "./types";
 
 import { FetchService } from "../fetch/service";
 import { GithubService } from "./service";
-import { githubUserMock } from "../../test/mocks";
 import { mock } from "jest-mock-extended";
+import { githubUserMock } from "../test/mocks";
 
 describe("GithubService", () => {
   const githubQuery: GeneralGithubQuery = {
