+++ api/src/test/mocks.ts
@@ -1,0 +1,25 @@
+import { GithubUser } from "../app/types/legacy";
+
+export const githubUserMock: GithubUser = {
+  avatar_url: "avatar_url",  
+  html_url: "html_url",  
+  id: 1,
+  login: "login",
+  type: "type",
+};
+
+export const githubUserMock2: GithubUser = {
+  avatar_url: "avatar_url2",  
+  html_url: "html_url2",  
+  id: 2,
+  login: "login2",
+  type: "type2",
+};
+
+export const githubUserMock3: GithubUser = {
+  avatar_url: "avatar_url3",  
+  html_url: "html_url3",  
+  id: 3,
+  login: "login3",
+  type: "type3",
+};
