--- api/test/mocks.ts
@@ -1,25 +1,0 @@
-import { GithubUser } from "../src/_common/types";
-
-export const githubUserMock: GithubUser = {
-  avatar_url: "avatar_url", // eslint-disable-line camelcase
-  html_url: "html_url", // eslint-disable-line camelcase
-  id: 1,
-  login: "login",
-  type: "type",
-};
-
-export const githubUserMock2: GithubUser = {
-  avatar_url: "avatar_url2", // eslint-disable-line camelcase
-  html_url: "html_url2", // eslint-disable-line camelcase
-  id: 2,
-  login: "login2",
-  type: "type2",
-};
-
-export const githubUserMock3: GithubUser = {
-  avatar_url: "avatar_url3", // eslint-disable-line camelcase
-  html_url: "html_url3", // eslint-disable-line camelcase
-  id: 3,
-  login: "login3",
-  type: "type3",
-};
