+++ api/tsconfig.json
@@ -1,20 +1,7 @@
 {
-  "$schema": "http://json.schemastore.org/tsconfig",
+  "extends": "@dzcode.io/tooling/tsconfig.json",
   "compilerOptions": {
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
-    "strict": true,
-    "incremental": true,
-    "lib": ["es2018", "esnext.asynciterable"],
-    "module": "commonjs",
-    "esModuleInterop": true,
-    "moduleResolution": "node",
-    "target": "es2017",
-    "sourceMap": true,
-    "declaration": true,
     "outDir": "dist"
   },
-  "include": ["src/**/*"]
+  "include": ["src"]
 }
