+++ lerna.json
@@ -1,5 +1,5 @@
 {
-  "packages": ["packages/*", "api", "data", "web", "mobile"],
+  "packages": ["packages/*", "api", "data"],
   "version": "0.0.0",
   "npmClient": "yarn",
   "useWorkspaces": true
