+++ package.json
@@ -3,10 +3,6 @@
   "devDependencies": {
     "husky": "^7.0.4",
     "lerna": "^4.0.0",
-    "lint-staged": "^11.2.3",
-    "npm-run-all": "^4.1.5",
-    "rimraf": "^3.0.2",
-    "sync-directory": "^4.0.5",
     "syncpack": "^5.8.15"
   },
   "lint-staged": {
@@ -16,38 +12,18 @@
   },
   "private": true,
   "scripts": {
-    "api:build": "yarn --cwd api build",
-    "api:build:watch": "yarn --cwd api build:watch",
-    "api:dev": "yarn --cwd api dev",
-    "api:sync": "yarn --cwd api common:sync",
-    "api:sync:watch": "yarn api:sync --watch",
-    "clean": "yarn workspaces run clean && rimraf ./node_modules",
-    "data:build": "yarn --cwd data build",
-    "data:build:watch": "yarn --cwd data build:watch",
-    "data:dev": "yarn --cwd data dev",
-    "data:sync": "yarn --cwd data common:sync",
-    "data:sync:watch": "yarn data:sync --watch",
+    "build": "yarn lerna run build:alone --stream",
+    "clean": "yarn lerna run clean:alone --stream",
     "dev": "echo \"Please run one of these commands:\\n\\nyarn dev:web\\n\\nyarn dev:mobile\\n\\nyarn dev:all\n\"",
-    "dev:all": "yarn && npm-run-all data:sync api:sync -p data:build api:build -p web:sync:watch web:dev mobile:sync:watch mobile:dev data:sync:watch data:dev api:sync:watch api:dev data:build:watch api:build:watch",
-    "dev:mobile": "yarn && npm-run-all data:sync api:sync -p data:build api:build -p mobile:sync:watch mobile:dev data:sync:watch data:dev api:sync:watch api:dev data:build:watch api:build:watch",
-    "dev:web": "yarn && npm-run-all data:sync api:sync -p data:build api:build -p web:sync:watch web:dev data:sync:watch data:dev api:sync:watch api:dev data:build:watch api:build:watch",
-    "lint:check": "yarn workspaces run lint:check",
-    "lint:fix": "yarn workspaces run lint:fix",
-    "mobile:dev": "yarn --cwd mobile dev",
-    "mobile:sync": "yarn --cwd mobile common:sync",
-    "mobile:sync:watch": "yarn mobile:sync --watch",
+    "lint": "yarn build && yarn lint:alone",
+    "lint:alone": "yarn lerna run lint:alone --stream",
     "postinstall": "(husky install && husky set .husky/pre-commit \"yarn workspaces run lint-staged && yarn lint-staged\") || exit 0",
-    "test": "yarn workspaces run test",
-    "test:cov": "yarn workspaces run test:cov",
-    "web:dev": "yarn --cwd web dev",
-    "web:sync": "yarn --cwd web common:sync",
-    "web:sync:watch": "yarn web:sync --watch"
+    "test": "yarn build && yarn test:alone",
+    "test:alone": "yarn lerna run test:alone --stream"
   },
   "workspaces": [
     "packages/*",
     "api",
-    "data",
-    "web",
-    "mobile"
+    "data"
   ]
 }
