+++ packages/utils/src/config/environment.ts
@@ -1,0 +1,3 @@
+export const environments = ["development", "staging", "production"] as const;
+
+export type Environment = typeof environments[number];
